added param & test
