+++ lib/index.js
@@ -189,7 +189,11 @@ const exprGen = {
     return res
   },
   item: function (node) { // [ ],
-    return `${node.type}.${node.name}`
+    let item = node.type
+    if (node.name) {
+      item += `.${node.name}`
+    }
+    return item
   },
   literal: function (node) { // [ ],
     return node.value
@@ -238,7 +242,11 @@ const exprGen = {
     return '(nop)'
   },
   param: function (node) { // ['items' ],
-    return '(param)'
+    let param = ''
+    node.items.forEach((item) => {
+      param = `${param}${exprGen[item.kind](item)}`
+    })
+    return `(param ${param})`
   },
   relop: function (node) { // ['left', 'right' ],
     return `(${node.type}.${node.operator} ${exprGen[node.left.kind](node.left)}` +
